requirements.txt
meshtastic
openai
pypubsub
